telegram_analyzer/requirements.txt
telethon
fastapi
uvicorn
sentence-transformers
scikit-learn
numpy
jinja2
python-dotenv
sqlalchemy
uvloop; sys_platform != 'win32'
websockets
vaderSentiment
